Remove Router and change Responder
