+++ src/main/java/com/pwdd/httpServer/Server.java
@@ -7,11 +7,11 @@ class Server {
   private ServerSocket serverSocket;
   private Socket socket;
   private String rootDirectory;
-  Router router;
+  Responder responder;
 
   Server(String _rootDirectory) {
     this.rootDirectory = _rootDirectory;
-    this.router = new Router(rootDirectory);
+    this.responder = new Responder(rootDirectory);
   }
 
   void listenAt(int portNumber) {
@@ -37,7 +37,7 @@ class Server {
     while(true) {
       try {
         openConnection();
-        new Thread(new ConnectionHandler(socket, router)).start();
+        new Thread(new ConnectionHandler(socket, responder)).start();
       } catch (Exception e) {
         e.printStackTrace();
         return;
